allow open_port to set env vars containing a trailing '=' character
The erlang:open_port spawn and spawn_executable directives can include
an {env, Env} directive to set up environment variables for the
spawned process. A bug in ert/emulator/sys/unix/sys.c prevented
applications from using {env, Env} to set an environment variable
whose value ended with a '=' (equal sign) character; the code mistook
the trailing equal sign as an indication that an environment variable
was to be cleared from the environment of the spawned process.

For example, passing an {env, Env} of

{env, [{"foo", "bar="}]}

would result in the code in sys.c seeing a string of the form

"foo=bar="

The code would see the final '=' character and assume the directive
wanted to clear a variable named "foo=bar" from the environment of the
spawned process, rather than seeing it as a directive to set the
environment variable "foo" to the value "bar=".

Fix this problem and add a new regression test for it to the port test
suite.